partial abstraction
