+++ packages/next-s3-upload/src/pages/api/s3-upload.ts
@@ -1,14 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import {
-  STSClient,
-  GetFederationTokenCommand,
-  STSClientConfig,
-} from '@aws-sdk/client-sts';
-import { PutObjectCommand } from '@aws-sdk/client-s3';
-import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
-import { getConfig, S3Config } from '../../utils/config';
-import { getClient } from '../../utils/client';
-import { sanitizeKey, uuid } from '../../utils/keys';
+import { S3Config } from '../../utils/config';
+import { route } from '../../utils/route-builder';
 
 type NextRouteHandler = (
   req: NextApiRequest,
@@ -23,100 +15,18 @@ type Options = S3Config & {
 };
 
 let makeRouteHandler = (options: Options = {}): Handler => {
-  let route: NextRouteHandler = async function(req, res) {
-    let config = getConfig({
-      accessKeyId: options.accessKeyId,
-      secretAccessKey: options.secretAccessKey,
-      bucket: options.bucket,
-      region: options.region,
-      forcePathStyle: options.forcePathStyle,
-      endpoint: options.endpoint,
+  const nextPageRoute: NextRouteHandler = async function(req, res) {
+    const { status, body } = await route({
+      body: req.body,
+      options: options, // Needs to be abstracted into RouteOptions still
     });
 
-    let missing = missingEnvs(config);
-    if (missing.length > 0) {
-      res
-        .status(500)
-        .json({ error: `Next S3 Upload: Missing ENVs ${missing.join(', ')}` });
-    } else {
-      let uploadType = req.body._nextS3?.strategy;
-      let filename = req.body.filename;
-
-      let key = options.key
-        ? await Promise.resolve(options.key(req, filename))
-        : `next-s3-uploads/${uuid()}/${sanitizeKey(filename)}`;
-      let { bucket, region, endpoint } = config;
-
-      if (uploadType === 'presigned') {
-        let filetype = req.body.filetype;
-        let client = getClient(config);
-        let params = {
-          Bucket: bucket,
-          Key: key,
-          ContentType: filetype,
-          CacheControl: 'max-age=630720000',
-        };
-
-        const url = await getSignedUrl(client, new PutObjectCommand(params), {
-          expiresIn: 60 * 60,
-        });
-
-        res.status(200).json({
-          key,
-          bucket,
-          region,
-          endpoint,
-          url,
-        });
-      } else {
-        let stsConfig: STSClientConfig = {
-          credentials: {
-            accessKeyId: config.accessKeyId,
-            secretAccessKey: config.secretAccessKey,
-          },
-          region,
-        };
-
-        let policy = {
-          Statement: [
-            {
-              Sid: 'Stmt1S3UploadAssets',
-              Effect: 'Allow',
-              Action: ['s3:PutObject'],
-              Resource: [`arn:aws:s3:::${bucket}/${key}`],
-            },
-          ],
-        };
-
-        let sts = new STSClient(stsConfig);
-
-        let command = new GetFederationTokenCommand({
-          Name: 'S3UploadWebToken',
-          Policy: JSON.stringify(policy),
-          DurationSeconds: 60 * 60, // 1 hour
-        });
-
-        let token = await sts.send(command);
-
-        res.status(200).json({
-          token,
-          key,
-          bucket,
-          region,
-        });
-      }
-    }
+    res.status(status).json(body);
   };
 
   let configure = (options: Options) => makeRouteHandler(options);
 
-  return Object.assign(route, { configure });
-};
-
-let missingEnvs = (config: Record<string, any>): string[] => {
-  let required = ['accessKeyId', 'secretAccessKey', 'bucket', 'region'];
-
-  return required.filter(key => !config[key] || config.key === '');
+  return Object.assign(nextPageRoute, { configure });
 };
 
 let APIRoute = makeRouteHandler();
